\name{cmmb_demo_data}
\alias{cmmb_demo_data}
\docType{data}
\title{
Example Data
}
\description{
It is a list that contains artificial 50 samples with a binary outcome variable y, a binary treatment tr, a compositional mediator M with 5 components, and 1 covariate X. The true direct and indirect effects of treatment on the outcome are 0.26 and 0.14, respectively.
}
\usage{data(cmmb_demo_data)}
\keyword{datasets}
